pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-airports-xtd"
version = "0.3.0"
description = "Django app providing airport data from OpenFlights"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Antonio Ercole De Luca", email = "eracle@posteo.eu"},
]
maintainers = [
    {name = "Ales Hanson"},
]
keywords = ["django", "airports", "geodjango", "openflights"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Scientific/Engineering :: GIS",
]
requires-python = ">=3.10"
dependencies = [
    "Django>=5.0",
    "django-cities-xtd>=0.7.0",
    "django-model-utils>=5.0.0",
    "tqdm>=4.0.0",
    "requests>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "ruff>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/arthanson/django-airports-xtd"
Documentation = "https://arthanson.github.io/django-airports-xtd/"
Repository = "https://github.com/arthanson/django-airports-xtd"
Issues = "https://github.com/arthanson/django-airports-xtd/issues"

[tool.setuptools.packages.find]
include = ["airports*"]

[tool.setuptools.package-data]
airports = [
    "static/**/*",
    "templates/**/*",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["airports"]
known-third-party = ["django", "cities"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_app.settings"
pythonpath = [".", "test_project"]
testpaths = ["test_project/test_app/tests"]
